:goat: Modify shuffle function to not mutate data & fix slice(1) bug
It is a React best practice to not mutate data used as props or state. The result variable is used to store the copy of list, however slice(1) was copying without index 0. Change slice(1) to slice(0) to fix this. Inorder to not mutate the list data, the result variable must also be used in the for loop when performing the shuffling.

 - Modify slice(1) to slice(0) to include index 0
 - Use result variable in shuffle for loop for swap